+++ portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/addcontactcc/AddContactCCTests.html
@@ -9,6 +9,7 @@
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>AddContactCCTests</b></td></tr>
 <tr><td><a href="AddContactCCTest.html">AddContactCCTest</a></td></tr>
+<tr><td><a href="ViewContactCCTest.html">ViewContactCCTest</a></td></tr>
 <tr><td><a href="TearDownContactTest.html">TearDownContactTest</a></td></tr>
 </tbody></table>
 </body>
