+++ portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/addcontactcc/AddContactCCTests.java
@@ -26,6 +26,7 @@ public class AddContactCCTests extends BaseTestSuite {
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
 		testSuite.addTestSuite(AddContactCCTest.class);
+		testSuite.addTestSuite(ViewContactCCTest.class);
 		testSuite.addTestSuite(TearDownContactTest.class);
 
 		return testSuite;
