--- portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/viewcontactcc/AddContactCCTest.html
@@ -1,78 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>AddContactCCTest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">AddContactCCTest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/user/joebloggs/so/dashboard/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//nav/ul/li[contains(.,'Contacts Center')]/a/span</td>
-	<td>Contacts Center</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//nav/ul/li[contains(.,'Contacts Center')]/a/span</td>
-	<td>Contacts Center</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//span[@class='add-contact']/a</td>
-	<td>Add</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//span[@class='add-contact']/a</td>
-	<td>Add</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//input[@id='_1_WAR_contactsportlet_fullName']</td>
-	<td>Name</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_fullName']</td>
-	<td>Social01 Office01 Contact01</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_emailAddress']</td>
-	<td>socialofficecontact01@liferay.com</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//textarea[@id='_1_WAR_contactsportlet_comments']</td>
-	<td>Social01 Office01 Contact01 Comments</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//input[@value='Save']</td>
-	<td>Save</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//span[@class='portlet-msg-success']</td>
-	<td>You have successfully added a new contact.</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//span[@class='portlet-msg-success']</td>
-	<td>You have successfully added a new contact.</td>
-</tr>
-</tbody></table>
-</body>
-</html>
