--- portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/viewcontactcc/TearDownContactTest.html
@@ -1,308 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>TearDownContactTest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">TearDownContactTest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/user/joebloggs/so/dashboard/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//nav/ul/li[contains(.,'Contacts Center')]/a/span</td>
-	<td>Contacts Center</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//nav/ul/li[contains(.,'Contacts Center')]/a/span</td>
-	<td>Contacts Center</td>
-</tr>
-<tr>
-	<td>assertVisible</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>Search</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>social</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>contact1Present</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${contact1Present}</td>
-	<td>Contact1NotPresent</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>Contact</td>
-</tr>
-<tr>
-	<td>storeText</td>
-	<td>//div[@class='lfr-contact-name']/a</td>
-	<td>contact1Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>clickAndWait</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to delete ${contact1Name} from your contacts?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>Contact1NotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>social</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>contact2Present</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${contact2Present}</td>
-	<td>Contact2NotPresent</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>Contact</td>
-</tr>
-<tr>
-	<td>storeText</td>
-	<td>//div[@class='lfr-contact-name']/a</td>
-	<td>contact2Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>clickAndWait</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to delete ${contact2Name} from your contacts?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>Contact2NotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>social</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>contact3Present</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${contact3Present}</td>
-	<td>Contact3NotPresent</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>Contact</td>
-</tr>
-<tr>
-	<td>storeText</td>
-	<td>//div[@class='lfr-contact-name']/a</td>
-	<td>contact3Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>clickAndWait</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to delete ${contact3Name} from your contacts?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>Contact3NotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>social</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>contact4Present</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${contact4Present}</td>
-	<td>Contact4NotPresent</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>Contact</td>
-</tr>
-<tr>
-	<td>storeText</td>
-	<td>//div[@class='lfr-contact-name']/a</td>
-	<td>contact4Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>clickAndWait</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to delete ${contact4Name} from your contacts?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>Contact4NotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_1_WAR_contactsportlet_name']</td>
-	<td>social</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>contact5Present</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${contact5Present}</td>
-	<td>Contact5NotPresent</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//div[@class='lfr-contact-info']</td>
-	<td>Contact</td>
-</tr>
-<tr>
-	<td>storeText</td>
-	<td>//div[@class='lfr-contact-name']/a</td>
-	<td>contact5Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>clickAndWait</td>
-	<td>//button[contains(.,'Delete')]</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to delete ${contact5Name} from your contacts?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>Contact5NotPresent</td>
-	<td></td>
-</tr>
-</tbody></table>
-</body>
-</html>
