--- portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/viewcontactcc/ViewContactCCTest.java
@@ -1,85 +1,0 @@
-/**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portalweb.socialofficehome.contactscenter.contacts.viewcontactcc;
-
-import com.liferay.portalweb.portal.BaseTestCase;
-import com.liferay.portalweb.portal.util.RuntimeVariables;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class ViewContactCCTest extends BaseTestCase {
-	public void testViewContactCC() throws Exception {
-		selenium.open("/user/joebloggs/so/dashboard/");
-		loadRequiredJavaScriptModules();
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible(
-							"//nav/ul/li[contains(.,'Contacts Center')]/a/span")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		selenium.clickAt("//nav/ul/li[contains(.,'Contacts Center')]/a/span",
-			RuntimeVariables.replace("Contacts Center"));
-		selenium.waitForPageToLoad("30000");
-		loadRequiredJavaScriptModules();
-		assertTrue(selenium.isVisible(
-				"//input[@id='_1_WAR_contactsportlet_name']"));
-		selenium.type("//input[@id='_1_WAR_contactsportlet_name']",
-			RuntimeVariables.replace("social"));
-		Thread.sleep(5000);
-		assertEquals(RuntimeVariables.replace("Social01 Office01 Contact01"),
-			selenium.getText("//div[@class='lfr-contact-name']"));
-		selenium.clickAt("//div[@class='lfr-contact-name']",
-			RuntimeVariables.replace("Social01 Office01 Contact01"));
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible(
-							"xPath=(//div[@class='lfr-contact-name'])[2]")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		assertEquals(RuntimeVariables.replace("Social01 Office01 Contact01"),
-			selenium.getText("xPath=(//div[@class='lfr-contact-name'])[2]"));
-		assertEquals(RuntimeVariables.replace(
-				"socialofficecontact01@liferay.com"),
-			selenium.getText("xPath=(//div[@class='lfr-contact-extra'])[2]"));
-		assertEquals(RuntimeVariables.replace(
-				"Social01 Office01 Contact01 Comments"),
-			selenium.getText("//div[@class='comments']"));
-	}
-}
