--- portal-web/test/com/liferay/portalweb/socialofficehome/contactscenter/contacts/viewcontactcc/ViewContactCCTests.html
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-  <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ViewContactCCTests</title>
-</head>
-<body>
-<table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ViewContactCCTests</b></td></tr>
-<tr><td><a href="AddContactCCTest.html">AddContactCCTest</a></td></tr>
-<tr><td><a href="ViewContactCCTest.html">ViewContactCCTest</a></td></tr>
-<tr><td><a href="TearDownContactTest.html">TearDownContactTest</a></td></tr>
-</tbody></table>
-</body>
-</html>
